+++ src/main/java/ds/FloatKeyedMaxHeap.java
@@ -33,7 +33,7 @@ public class FloatKeyedMaxHeap<T> extends AbstractFloatKeyedArrayHeap<T> {
     }
 
     @Override
-    public void increaseElementKeyBy(int index, int offset) {
+    public void increaseElementKeyBy(int index, float offset) {
         this.elements[index].increaseKeyBy(offset);
         if (offset == 0){
             return;
