+++ src/main/java/ds/FloatKeyedMinHeap.java
@@ -33,7 +33,7 @@ public class FloatKeyedMinHeap<T> extends AbstractFloatKeyedArrayHeap<T> {
     }
 
     @Override
-    public void increaseElementKeyBy(int index, int offset) {
+    public void increaseElementKeyBy(int index, float offset) {
         this.elements[index].increaseKeyBy(offset);
         if (offset == 0){
             return;
